pyproject.toml
[project]
name = "chat-llm-v3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
  "langchain>=1.2.0",
  "langgraph>=1.0.5",
  "langchain-ollama>=1.0.1",
  "chromadb>=1.4.0",
  "requests>=2.31.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.0",
  "rich>=13.7.0",
  "firecrawl-py>=4.12.0",
  "fastapi>=0.128.0",
  "langchain-google-genai>=4.1.2",
]

[tool.uv]
package = true

[tool.setuptools]
packages = ["app"]

[tool.pyright]
venvPath = "."
venv = ".venv"
